# babson-scheduler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Babson Library Student Worker Scheduler",
    "description": "\n    Constraint-aware scheduling system for Babson Library student workers.\n\n    Uses Google OR-Tools CP-SAT for optimization-based schedule generation.\n    Designed for IBM Orchestrate integration via OpenAPI custom extension.\n    ",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://your-deployment-host.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/availability/": {
      "post": {
        "tags": [
          "Availability"
        ],
        "summary": "Submit Availability",
        "operationId": "submit_availability_api_v1_availability__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilitySubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Submit Availability Api V1 Availability  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Availability"
        ],
        "summary": "List Availability",
        "operationId": "list_availability_api_v1_availability__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "week_start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailabilitySlotOut"
                  },
                  "title": "Response List Availability Api V1 Availability  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/availability/heatmap/{week_start_date}": {
      "get": {
        "tags": [
          "Availability"
        ],
        "summary": "Get Heatmap",
        "operationId": "get_heatmap_api_v1_availability_heatmap__week_start_date__get",
        "parameters": [
          {
            "name": "week_start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Week Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoverageHeatmapEntry"
                  },
                  "title": "Response Get Heatmap Api V1 Availability Heatmap  Week Start Date  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/availability/import": {
      "post": {
        "tags": [
          "Availability"
        ],
        "summary": "Import Availability",
        "operationId": "import_availability_api_v1_availability_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_availability_api_v1_availability_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Import Availability Api V1 Availability Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilitySlot": {
        "properties": {
          "shift_instance_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Instance Id",
            "description": "Preferred: link directly to shift_instance. Provide this OR (day_of_week + times)."
          },
          "day_of_week": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week",
            "description": "0=Monday \u2026 6=Sunday"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time",
            "description": "HH:MM"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time",
            "description": "HH:MM"
          },
          "level": {
            "type": "string",
            "enum": [
              "preferred",
              "available",
              "cannot_work"
            ],
            "title": "Level",
            "description": "Availability level for this slot"
          }
        },
        "type": "object",
        "required": [
          "level"
        ],
        "title": "AvailabilitySlot"
      },
      "AvailabilitySlotOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "week_start_date": {
            "type": "string",
            "title": "Week Start Date"
          },
          "shift_instance_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Instance Id"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "import_source": {
            "type": "string",
            "title": "Import Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "week_start_date",
          "shift_instance_id",
          "day_of_week",
          "start_time",
          "end_time",
          "level",
          "submitted_at",
          "import_source"
        ],
        "title": "AvailabilitySlotOut"
      },
      "AvailabilitySubmitRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id",
            "description": "UUID of the submitting student"
          },
          "week_start_date": {
            "type": "string",
            "format": "date",
            "title": "Week Start Date",
            "description": "Monday of the target week"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/AvailabilitySlot"
            },
            "type": "array",
            "title": "Slots",
            "description": "All availability slots. Atomically replaces any previous submission for this student+week."
          },
          "import_source": {
            "type": "string",
            "enum": [
              "api",
              "excel",
              "form"
            ],
            "title": "Import Source",
            "default": "api"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "week_start_date",
          "slots"
        ],
        "title": "AvailabilitySubmitRequest"
      },
      "Body_import_availability_api_v1_availability_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "week_start_date": {
            "type": "string",
            "format": "date",
            "title": "Week Start Date"
          },
          "student_id_column": {
            "type": "string",
            "title": "Student Id Column",
            "default": "name"
          }
        },
        "type": "object",
        "required": [
          "file",
          "week_start_date"
        ],
        "title": "Body_import_availability_api_v1_availability_import_post"
      },
      "CoverageHeatmapEntry": {
        "properties": {
          "shift_instance_id": {
            "type": "string",
            "title": "Shift Instance Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "slot_index": {
            "type": "integer",
            "title": "Slot Index"
          },
          "is_hard_shift": {
            "type": "boolean",
            "title": "Is Hard Shift"
          },
          "preferred_count": {
            "type": "integer",
            "title": "Preferred Count"
          },
          "available_count": {
            "type": "integer",
            "title": "Available Count"
          },
          "total_eligible": {
            "type": "integer",
            "title": "Total Eligible"
          },
          "coverage_risk": {
            "type": "string",
            "enum": [
              "none",
              "low",
              "critical"
            ],
            "title": "Coverage Risk",
            "description": "none=3+ eligible, low=1-2, critical=0"
          }
        },
        "type": "object",
        "required": [
          "shift_instance_id",
          "date",
          "day_of_week",
          "start_time",
          "end_time",
          "slot_index",
          "is_hard_shift",
          "preferred_count",
          "available_count",
          "total_eligible",
          "coverage_risk"
        ],
        "title": "CoverageHeatmapEntry"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}